+++ src/charging-station/ui-server/UIServerAccessPolicy.ts
@@ -3,14 +3,15 @@ import type { IncomingMessage } from 'node:http'
 import type { UIServerConfiguration } from '../../types/index.js'
 
 import { UI_SERVER_ACCESS_POLICY_DEFAULTS } from '../../utils/ConfigurationSchema.js'
-import { has, isEmpty, isNotEmptyArray } from '../../utils/index.js'
 import {
+  has,
+  isEmpty,
   isLoopback,
+  isNotEmptyArray,
   normalizeHost,
   normalizeIPAddress,
-  splitHeaderList,
-  splitQuoted,
-} from './UIServerNet.js'
+} from '../../utils/index.js'
+import { splitHeaderList, splitQuoted } from './UIServerNet.js'
 
 const FORWARDED_HEADER_NAMES = [
   'forwarded',
