+++ src/charging-station/ui-server/UIServerFactory.ts
@@ -9,10 +9,9 @@ import {
   ConfigurationSection,
   type UIServerConfiguration,
 } from '../../types/index.js'
-import { logger, logPrefix } from '../../utils/index.js'
+import { isLoopback, logger, logPrefix } from '../../utils/index.js'
 import { UIHttpServer } from './UIHttpServer.js'
 import { UIMCPServer } from './UIMCPServer.js'
-import { isLoopback } from './UIServerNet.js'
 import { UIWebSocketServer } from './UIWebSocketServer.js'
 
 const moduleName = 'UIServerFactory'
